tests: Add bundle test deployment for EKS
Add integration test that deploys the bundle and asserts that it is
deployed without errors and is reachable through its public endpoint.
This test was written with the purpose of being used for deployment to
EKS and uses a public url that extracts from the service
istio-ingressgateway-workload's .status.loadBalancer.ingress[0].hostname
field. Generalizing this will allow us to use this test for other
deployemtns as well.

Signed-off-by: Orfeas Kourkakis <orfeas.kourkakis@canonical.com>